Use decimal literals for decimal parser expectations (#344)

+++ test/Parlot.Tests/FluentTests.cs
@@ -471,7 +471,7 @@ public class FluentTests
             }, parsers);
 
         Assert.True(parser.TryParse("d:123.456", out var resultD));
-        Assert.Equal((decimal)123.456, resultD);
+        Assert.Equal(123.456m, resultD);
 
         Assert.True(parser.TryParse("i:123", out var resultI));
         Assert.Equal((long)123, resultI);
@@ -1350,7 +1350,7 @@ public class FluentTests
     [Fact]
     public void NumberParsesCustomDecimalSeparator()
     {
-        Assert.Equal((decimal)123.456, Literals.Number<decimal>(NumberOptions.Any, decimalSeparator: '|').Parse("123|456"));
+        Assert.Equal(123.456m, Literals.Number<decimal>(NumberOptions.Any, decimalSeparator: '|').Parse("123|456"));
     }
 
     [Fact]
